typeformx v0.0.1 Typeform.ClientField

Summary module contains the struct for the Client Form Field from the API - used for creating a form fields

Functions

build(data)

builds a ClientField struct based on a map that has the correct variables

valid_field?(arg1)